pyproject.toml
[tool.poetry]
name = "wsidicom"
version = "0.16.0"
description = "Tools for handling DICOM based whole scan images"
authors = ["Erik O Gabrielsson <erik.o.gabrielsson@sectra.com>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/imi-bigpicture/wsidicom"
packages = [{include = "wsidicom"}]
keywords = ["whole slide image", "digital pathology", "annotations", "dicom"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Medical Science Apps."
]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.22.0"
pydicom = ">=2.1.0"
Pillow = ">=9.1.1"
dicomweb-client = "^0.59.1"
imagecodecs = { version = "^2023.9.18", optional = true }
pylibjpeg-rle = { version = "^1.3.0", optional = true }

[tool.poetry.extras]
imagecodecs = ["imagecodecs"]
rle = ["pylibjpeg-rle"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-watch = "^4.2.0"
xmltodict = "^0.12.0"
shapely = "^2.0.2"
pycodestyle = "^2.8.0"
black = "^23.1.0"
flake8 = "^4.0.1"
codespell = "^2.2.5"
wsidicom-data =  "^0.3.0"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
